+++ src/CodeActionUtil.ts
@@ -24,6 +24,10 @@ export class CodeActionUtil {
                 edit.changes![uri].push(
                     TextEdit.replace(change.range, change.newText)
                 );
+            } else if (change.type === 'delete') {
+                edit.changes![uri].push(
+                    TextEdit.del(change.range)
+                );
             }
         }
         const action = CodeAction.create(obj.title, edit, obj.kind);
@@ -51,7 +55,7 @@ export interface CodeActionShorthand {
     diagnostics?: BsDiagnostic[];
     kind?: CodeActionKind;
     isPreferred?: boolean;
-    changes: Array<InsertChange | ReplaceChange>;
+    changes: Array<InsertChange | ReplaceChange | DeleteChange>;
 }
 
 export interface InsertChange {
@@ -67,5 +71,12 @@ export interface ReplaceChange {
     type: 'replace';
     range: Range;
 }
+
+export interface DeleteChange {
+    filePath: string;
+    type: 'delete';
+    range: Range;
+}
+
 
 export const codeActionUtil = new CodeActionUtil();